Separated the controllers.
