+++ Rakefile
@@ -1,7 +1,7 @@
 require 'data_mapper'
 require './app/app.rb'
 if ENV['RACK_ENV'] != 'production'
-  require 'rspec/core/rake_task'
+require 'rspec/core/rake_task'
 
   RSpec::Core::RakeTask.new :spec
