+++ app/controllers/peep.rb
@@ -1,0 +1,22 @@
+class Chitter < Sinatra::Base
+  get '/peeps/new' do
+    erb :new_peep
+  end
+
+  post '/peeps' do
+    if session[:user_id] != nil
+      current_user.peeps << Peep.create(body: params[:peep], name: current_user.name, username: current_user.username)
+      @peep = Peep.all
+      current_user.save
+      erb :show_peep
+    else
+      flash[:errors] = ['You are not logged in']
+      redirect '/peeps/new'
+    end
+
+  end
+
+  get '/all_peeps' do
+    erb :show_all_peeps
+  end
+end
