+++ app/models/datamapper_settings.rb
@@ -1,9 +1,7 @@
 require 'data_mapper'
-
-def data_mapper_settings
+require_relative './peep'
+require_relative './user'
 
   DataMapper.setup(:default, ENV['DATABASE_URL'] || "postgres://localhost/chitter_#{ENV['RACK_ENV']}")
   DataMapper.finalize
   DataMapper.auto_upgrade!
-
-end
