+++ spec/features/failing_signup_spec.rb
@@ -1,5 +1,4 @@
 require 'spec_helper'
-require_relative '../web-helper'
 
 feature "Failing Sign Up" do
   include Helpers
